test(workspace.rs): test New Floating Window
Created a test to test adding a new floating window. The test creates a container with three windows
and attempts to add one of the windows to the list of floating windows. the test checks to see if
the first window in the container was added to the floating window list and also checks the list of
windows and floating windows to ensure we get the expected result.