feat(bar): added locked_container widget, use accent colour for icons
This commit adds a new komorebi widget to indicate whether or not the focused container is locked.

This commit also includes an icon colour change on the layer and layout widgets to the accent colour.

The commit also renames the locked_window widget to locked_container as it is more suitable.

PR: #1394