feat(wm): opening the apply state function as a socketmessage for the first lower level command
this allows clients to send a state and get it is applied to the wm. but by using the already there
apply_state function we have all the safety we need to apply the sended state. this is also with
attention not open for the komorebic binary because I think the average user doesnt need such a
command and this is much safer also to not give them the ability. he question is if this should be a
seperated function or not