feat(bar): send commands by mouse/touchpad/screen
This commit makes it possible to send commands from the bar by using the
mouse/touchpad/touchscreen.

Komorebi or custom commands can be sent by clicking on the
mouse's primary, secondary, middle, back or forward buttons.

As the primary single click is already used by widgets, only primary double clicks can
send commands. This limitation is due to Egui also triggering 2 single clicks
before a double click is triggered. Egui does not have an
implementation for stopping event propagation out of the box and would
be too much work to include.

Similarly, commands can be sent on every "tick" of mouse scrolling, touchpad or
touchscreen swiping in any of the 4 directions. This "tick" can be
adjusted to fit user's preference.

This is due to the fact, that Egui does not have an event for when a mouse "tick" occurs. It instead
gives a number of points that the user scrolled/swiped on each frame.

PR: #1403