test(window_manager.rs): monocle on and off test
Created a test for testing the monocle_on and monocle_off functions. The test checks to ensure that
the focused workspace container becomes a monocle container when calling the monocle_on function.
The test will also ensure that the container is moved to the workspace when the monocle_off function
is called.