test(window_manager.rs): ensure named workspace for monitor test
Created a test that tests the ensure_named_workspace_for_monitor function. the test creates two
monitors and holds a list of workspace names. When calling the
ensure_named_workspace_for_monitor_function the test checks to ensure that the the the monitor
contains the length of the list workspaces and that the workspaces uses the name in the list. The
test adds more names to the list and repeats the check on the other monitor.