test(monitor.rs): test moving a container to a nonexistent workspace
Created a test that test moving a container to a workspace that doesn't exist. The test checks to
ensure when a container is moved to a workspace that doesn't exist, the workspace is created and
that the workspace contains to container. It also checks that there are n number of workspaces
available where n is the largest workspace number.