test(window_manager.rs): test transfer window to nonexistent monitor
Created a test that tests the transfer_window function. The tests attempts to transfer a window to a
monitor that doesn't exist. The test checks to see if we return an error. The test successfully gets
an error but there is a bug where the window isn't in the contiainer after a failed transfer. I
wrote a note comment to explain the bug just in case we need to reference back to it.