test(window_manager.rs): test swap container with non existent container
Created a test that tests swapping a container with a container that doesn't exist. The testensures
that we receive an error and that the contents of the container is still available when attempting
to swap a container with a nonexistent one. A bug was discovered where the contents of the container
was being removed even though the swap_container function would return an error. LGUGZ discovered
that the issue was due to the contents of the container being removed before verifying that both
containers exist. The fix that LGUGZ came up with to fix the bug is included.