feat: add the ability to set work-area-offset per workspace
This commit adds the option to set 'work_area_offset' per workspace. If no workspace work area
offset is set for that workspace it will instead use the value of the globals.work_area_offset for
that workspace.

This commit adds a command to set the work area offset of a workspace given a monitor index and a
workspace index.