fix(bar): update the network widget in a separate thread
This commit changes the way how the default interface and network activity is loaded by spawning a new thread on each request. This way the main thread is not blocked by this process.

There has been instances where getting the default interface blocked the egui ui thread, resulting in a frozen bar.

Fixes: #1423 #1499