feat(wm): allow scrolling layout with >1 monitors
This commit takes some inspiration from a trick I implemented in the
komorebi for Mac codebase to hide windows that are not within the
horizontal bounds of the work space area when the scrolling layer is
active.

Doing this ensures that such windows will not spill over into adjacent
monitors (or at least, not for very long).

There are some additional changes required in komorebi for Windows to
try and make sure this plays nicely with the animations feature, which
is not present in komorebi for Mac.